+++ example/lib/screens/home/home.screen.dart
@@ -69,7 +69,8 @@ class _HomeScreenState extends State<HomeScreen> {
                     shrinkWrap: true,
                     itemCount: snapshot.data!.docs.length,
                     itemBuilder: (context, index) {
-                      final category = CategoryModel.fromSnapshot(snapshot.data!.docs[index]);
+                      final category = CategoryModel.fromSnapshot(
+                          snapshot.data!.docs[index]);
                       return ListTile(
                         title: Text(category.title),
                         subtitle: Text('category: ${category.category}'),
@@ -96,7 +97,8 @@ class _HomeScreenState extends State<HomeScreen> {
               child: const Text('Delete all files on storage'),
             ),
             ElevatedButton(
-              onPressed: () => snackBarSuccess(title: 'title', message: 'message'),
+              onPressed: () =>
+                  snackBarSuccess(title: 'title', message: 'message'),
               child: const Text('snackbar'),
             ),
           ],