+++ lib/src/test/test.utils.dart
@@ -8,14 +8,16 @@ Future loadUsers() async {
   TestConfig.b = await UserService.instance.getByEmail(TestConfig.emailB);
   TestConfig.c = await UserService.instance.getByEmail(TestConfig.emailC);
   TestConfig.d = await UserService.instance.getByEmail(TestConfig.emailD);
-  TestConfig.admin = await UserService.instance.getByEmail(TestConfig.adminEmail);
+  TestConfig.admin =
+      await UserService.instance.getByEmail(TestConfig.adminEmail);
 }
 
 Future clear() async {
   dog('Clear all posts of test users');
   await loginAsAdmin();
   wait(500);
-  final snapshot = await PostService.instance.col.where('userDocumentReference', whereIn: [
+  final snapshot =
+      await PostService.instance.col.where('userDocumentReference', whereIn: [
     TestConfig.a.ref,
     TestConfig.b.ref,
     TestConfig.c.ref,
@@ -42,7 +44,8 @@ Future loginAsAdmin() async {
 
 Future loginAs(String email, [String? password]) async {
   dog('Login as $email');
-  await UserService.instance.loginOrRegister(email, password ?? TestConfig.password);
+  await UserService.instance
+      .loginOrRegister(email, password ?? TestConfig.password);
   await UserService.instance.publicRef.update({'email': email});
   dog('uid: ${UserService.instance.uid}');
 }
@@ -57,8 +60,10 @@ Future createPost() async {
   final ref = await PostService.instance.col.add({
     'category': 'qna',
     'userDocumentReference': UserService.instance.ref,
-    'title': 'Created by ${UserService.instance.my.data['email']} at ${DateTime.now()}',
-    'content': 'Content. Created by ${UserService.instance.my.data['email']} at ${DateTime.now()}',
+    'title':
+        'Created by ${UserService.instance.my.data['email']} at ${DateTime.now()}',
+    'content':
+        'Content. Created by ${UserService.instance.my.data['email']} at ${DateTime.now()}',
   });
   await PostService.instance.afterCreate(postDocumentReference: ref);
 }
@@ -66,7 +71,9 @@ Future createPost() async {
 /// Deletes all the posts of the login user.
 Future deletePosts() async {
   dog('Delete all posts');
-  final snapshot = await PostService.instance.col.where('userDocumentReference', isEqualTo: UserService.instance.ref).get();
+  final snapshot = await PostService.instance.col
+      .where('userDocumentReference', isEqualTo: UserService.instance.ref)
+      .get();
   for (final doc in snapshot.docs) {
     await doc.reference.delete();
   }